Create meshFor2x2x2Voxels
The following mesh was generated for a voxel grid of size 2 by 2 by 2.  Each voxel is a cube of side length 4mm. 

The output of the generated mesh can be interpreted according to details in the gmsh Reference Manual for Gmsh,
version 2.8, 9 July 2014
http://geuz.org/gmsh/doc/texinfo/gmsh.pdf
accessed 09 January 2015, Copyright 1997-2014 Christophe Geuzaine, Jean-Francois Remacle.